# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "小知，你的校园助手",
        "title": "KnowEase API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/{userid}/post/campus": {
            "get": {
                "description": "获取生活区的校园标签页面的帖子信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区-分区"
                ],
                "summary": "校园标签页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应信息以及帖子信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "校园帖子查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/food": {
            "get": {
                "description": "获取生活区的美食标签页面的帖子信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区-分区"
                ],
                "summary": "美食标签页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应信息以及帖子信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "美食帖子查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/life": {
            "get": {
                "description": "获取生活区的生活标签页面的帖子信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区-分区"
                ],
                "summary": "生活标签页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应信息以及帖子信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "生活帖子查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/paint": {
            "get": {
                "description": "获取生活区的绘画标签页面的帖子信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区-分区"
                ],
                "summary": "绘画标签页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应信息以及帖子信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "绘画帖子查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/publish": {
            "post": {
                "description": "用户发帖，上传帖子正文、图片等信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区-发帖"
                ],
                "summary": "用户发帖",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "发帖内容",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PostMessage"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "成功的响应信息以及帖子信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "帖子发布失败！",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/recommend": {
            "get": {
                "description": "获取生活区的推荐标签页面的帖子信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区-分区"
                ],
                "summary": "推荐标签页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应信息以及帖子信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "推荐帖子查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/{userid}/post/{postid}": {
            "get": {
                "description": "用户点开帖子详情页，获取帖子详情信息,帖子浏览量加一",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区-分区"
                ],
                "summary": "获取帖子详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应信息以及帖子信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/post/{postid}/deletecomment": {
            "delete": {
                "description": "用户在帖子下面发布评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区-评论"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论ID",
                        "name": "commentid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/post/{postid}/publishcomment": {
            "post": {
                "description": "用户在帖子下面发布评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区-评论"
                ],
                "summary": "发布评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "帖子ID",
                        "name": "postid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "成功响应信息以及评论信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "输入无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "发布失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/{userid}/post/{postid}/{commentid}/{replyid}": {
            "delete": {
                "description": "用户删除自己的回复",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "生活区-评论"
                ],
                "summary": "删除回复",
                "parameters": [
                    {
                        "type": "string",
                        "description": "回复ID",
                        "name": "replyid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "输入无效，请重试!",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.PostMessage": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "comment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Comment"
                    }
                },
                "createAt": {
                    "type": "string"
                },
                "likeCount": {
                    "type": "integer"
                },
                "postID": {
                    "type": "string"
                },
                "posterName": {
                    "type": "string"
                },
                "posterURL": {
                    "type": "string"
                },
                "posterid": {
                    "type": "string"
                },
                "saveCount": {
                    "type": "integer"
                },
                "tag": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "urls": {
                    "type": "string"
                },
                "viewCount": {
                    "type": "integer"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}